OEL-1169: Move tests to non-javascript. Adapt to the translations restrictions and the eTrans block has been moved.
